+++ oxe-api/test/resource/form/test_get_form_question_enums.py
@@ -12,5 +12,5 @@ class TestGetFormQuestionEnums(BaseCase):
         self.assertEqual(200, response.status_code)
         self.assertEqual({
             'status': ['ACTIVE', 'INACTIVE', 'DELETED'],
-            'type': ['TEXT', 'CHECKBOX', 'OPTIONS']
+            'type': ['TEXT', 'CHECKBOX', 'OPTIONS', 'SELECT']
         }, response.json)